requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bidict==0.23.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1744743065687/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
datasets==3.5.1
decorator==5.2.1
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
einops==0.8.1
eval_type_backport==0.2.2
executing==2.2.0
filelock==3.18.0
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1743732186120/work
frozenlist==1.6.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.25.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459255340/work
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistralai==1.8.1
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
munkres==1.1.4
networkx==3.4.2
nnsight==0.4.6
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1745118880749/work/dist/numpy-2.2.5-cp313-cp313-linux_x86_64.whl#sha256=b9870dc50b40c48dc68353df842c84f4a6c87e4b0e908e23fb21b12e207bee54
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.76.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929688808/work
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySide6==6.9.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==78.1.1
shiboken6==6.9.0
simple-websocket==1.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615881993/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
